+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Remove `component.BaseProcessorFactory`, use `processorhelper.NewFactory` instead (#4175)
 - Move `service/parserprovider` package to `config/configmapprovider` (#4206)
 - Remove `config.Pipeline.Name` (#4326)
+- Remove `config.Pipeline.InputDataType` (#4343)
 
 ## v0.38.0 Beta